refactor mobile scroller styles into own file
